PACT-985 | EPR | RDF4J added to docker-compose.yml
RDF4J added to docker compose. In addition to this, a new BulkLoadData script has been created to load the current test data stored in resources, and a repository config file has been created for an In Memory store for testing.

README.md updated to reflect the configuration steps to have this component running before running integration tests.